+++ Project1/Animation.h
@@ -7,7 +7,7 @@ public:
 	Animation(sf::Texture* texture,sf::Vector2u imageCount, float switchTime);
 	~Animation();
 
-	void Update(int row, int deltaTime,bool faceRight);
+	void Update(int row, float deltaTime,bool faceRight);
 
 public:
 	sf::IntRect uvRect;
